requirements.txt
telicent-lib==3.1.1 
telicent-ies-tool==1.0.4
telicent-label-builder==0.1.6
python-dotenv==1.0.1
setuptools
polars